Django-Middleware-0x03/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-filter
